+++ src/paillier/params.rs
@@ -9,7 +9,7 @@ use zeroize::Zeroize;
 
 use crate::{
     tools::hashing::Hashable,
-    uint::{HasWide, ToMontgomery},
+    uint::{Extendable, MulWide},
 };
 
 pub trait PaillierParams: core::fmt::Debug + PartialEq + Eq + Clone + Send + Sync {
@@ -27,8 +27,8 @@ pub trait PaillierParams: core::fmt::Debug + PartialEq + Eq + Clone + Send + Syn
         + RandomPrimeWithRng
         + Serialize
         + for<'de> Deserialize<'de>
-        + HasWide<Wide = Self::Uint>
-        + ToMontgomery
+        + MulWide<Self::HalfUint, Self::Uint>
+        + Extendable<Self::Uint>
         + Zeroize;
 
     /// A modulo-residue counterpart of `HalfUint`.
@@ -45,13 +45,13 @@ pub trait PaillierParams: core::fmt::Debug + PartialEq + Eq + Clone + Send + Syn
         + ConstantTimeGreater
         + Encoding<Repr: Zeroize>
         + Hashable
-        + HasWide<Wide = Self::WideUint>
+        + MulWide<Self::Uint, Self::WideUint>
+        + Extendable<Self::WideUint>
         + InvMod<Output = Self::Uint>
         + RandomMod
         + RandomPrimeWithRng
         + Serialize
         + for<'de> Deserialize<'de>
-        + ToMontgomery
         + Zeroize;
 
     /// A modulo-residue counterpart of `Uint`.
@@ -71,11 +71,11 @@ pub trait PaillierParams: core::fmt::Debug + PartialEq + Eq + Clone + Send + Syn
         + ConditionallySelectable
         + Encoding<Repr: Zeroize>
         + Hashable
-        + HasWide<Wide = Self::ExtraWideUint>
+        + MulWide<Self::WideUint, Self::ExtraWideUint>
+        + Extendable<Self::ExtraWideUint>
         + RandomMod
         + Serialize
         + for<'de> Deserialize<'de>
-        + ToMontgomery
         + Zeroize;
 
     /// A modulo-residue counterpart of `WideUint`.