+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bumped MSRV to 1.83 ([#176])
 - The API is now generic over the elliptic curve ([#186])
 - Added `k256` (Secp256k1 parameters), `bip32` (BIP32 support), and `dev` (`tiny-curve` parameters) features. ([#199])
+- Removed `HasWide` requirement from `PaillierParams` types, in favor of `Extendable` and `MulWide` with blanket impl for all `Uint`s. ([#205])
 
 
 [#156]: https://github.com/entropyxyz/synedrion/pull/156
@@ -20,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#176]: https://github.com/entropyxyz/synedrion/pull/176
 [#186]: https://github.com/entropyxyz/synedrion/pull/186
 [#199]: https://github.com/entropyxyz/synedrion/pull/199
+[#205]: https://github.com/entropyxyz/synedrion/pull/205
 
 
 ## [0.2.0] - 2024-11-17